Conflicts:
	e-Look/src/main/java/beans-config.xml

+++ e-Look/src/main/java/com/e_Look/reportMessage/model/ReportMessageService_JNDI.java
@@ -25,7 +25,7 @@ public class ReportMessageService_JNDI {
 		
 	}
 	
-	public String getJSON(Integer status) {
+	public String getJSON(Byte status) {
 		return dao.getJSON(status);
 	}
